fix: when running on dev mode, use aliases in src/ and not in dist/. Update package.json accordingly
TS aliases are managed by tsconfig.json, whereas JS aliases are managed by the module-alias package.
They conflict if ts-node evaluates require('module-alias/register')

fix #283
