+++ Dockerfile
@@ -11,7 +11,7 @@ RUN yarn install
 
 COPY . .
 
-RUN ./node_modules/typescript/bin/tsc -p .
+RUN yarn build
 
 HEALTHCHECK --interval=60s --timeout=2s --retries=3 CMD wget ${HOST}:${PORT}/healthz -q -O - > /dev/null 2>&1
